requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-Login>=0.6
Werkzeug==2.3.7
gunicorn
Flask-WTF==1.1.1

